Switch helper functions to returning a number
This hides their complexity (including complex return types) inside
themselves and makes makes the final `count()` function simpler.

As for the `count()` function, I considered adding a TypeScript ignore,
though the chance for adding a fourth type and not dealing with it here
is real. Also, the three `if` clauses aren't so hard to read.