Introduce a new, narrower, settings type
The idea came from @sirreal ([0], [1]) to use the already defined keys
in defaultSettings as a natural way to have a stronger type guarantee.

Once we have the narrower type, we can distinguish between settings
object we expect from the user, where everything is optional and a
settings object that comes from `loadSettings()` where we know that a
massive amount of the properties have been set as part of `defaultSettings`.

[0] https://git.io/JfVWj
[1] https://git.io/JfVle