Based on previous commit, relax undefined checks
Since we have a typecheck guarantee that most properties are there,
we don’t need to do all those checks.

Note that we still do the checks for `shortcodesRegExp` and `type`. Since
the guarantee for them doesn’t come from `defaultSettings`, but from
`loadSettings()`, we would have needed yet another layer of type to
describe the settings coming from `defaultSettings` + `type`, `shortcodes`,
and `shortcodesRegExp`.

Personally, this crossed a line, and I decided
that leaving the checks in `stripShortcodes.js` and the `default` in the
`switch` in `count()` were a smaller price to pay than adding yet another
type and dealing with type trickery (would’ve needed a `Pick` and to
disable `jsdoc/valid-types` again).